+++ Code/GraphMol/CIPLabeler/rules/Rule6.h
@@ -22,7 +22,7 @@ class Rule6 : public SequenceRule {
  public:
   Rule6();
 
-  int compare(const Edge *a, const Edge *b) const override;
+  int8_t compare(const Edge *a, const Edge *b) const override;
 };
 
 }  // namespace CIPLabeler
